adds nominate-experts feature
